requirements.txt
aiogram==2.22.2
requests
python-dotenv